feat: report관련 dto 테스트 작성
