+++ src/main/java/com/woowacourse/pelotonbackend/certification/presentation/dto/CertificationCreateRequest.java
@@ -7,6 +7,8 @@ import javax.validation.constraints.NotNull;
 
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
 
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.woowacourse.pelotonbackend.certification.domain.Certification;
 import com.woowacourse.pelotonbackend.certification.domain.CertificationStatus;
 import com.woowacourse.pelotonbackend.vo.ImageUrl;
@@ -15,10 +17,11 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-@AllArgsConstructor(access = AccessLevel.PUBLIC, onConstructor_ = @ConstructorProperties({"status", "description",
+@AllArgsConstructor(onConstructor_ = @ConstructorProperties({"status", "description",
     "riderId", "missionId"}))
 @Builder
 @Getter
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 public class CertificationCreateRequest {
     @NotNull
     private final CertificationStatus status;
