+++ src/test/java/com/woowacourse/pelotonbackend/infra/login/dto/KakaoTokenResponseTest.java
@@ -4,6 +4,7 @@ import static com.woowacourse.pelotonbackend.member.domain.LoginFixture.*;
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 import static org.junit.jupiter.api.Assertions.*;
 
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -12,6 +13,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 class KakaoTokenResponseTest {
     private final ObjectMapper objectMapper = new ObjectMapper();
 
+    @DisplayName("KakaoTokenResponse가 정상적으로 Serialize되는 지 확인한다.")
     @Test
     void JsonToResponse() throws JsonProcessingException {
         final String kakaoTokenResponseBody = "{\n"
